chore: create test cases for thunder client
If you are using thunder client you can import the .json file to create a collection.
With the collection created, create a env inside thunder client and link it to your .env file.
After this, you'll need to provide the base url for your tests.
When it is all set you can test the API with the collection by running all requests.