feat: create endpoint for sign-up
To create a user send a POST request to /sign-up.
On the body of the request there must be a email and password.
If you send the body incorrectly you'll receive a 422 status code with a joi error.
If the email is already in use you'll receive a 409.
If everything is ok you'll receive 201 and the id and email created.